Moved work package templates to their own file
