+++ src/backend/src/routes/work-package-templates.routes.ts
@@ -1,5 +1,4 @@
 import express from 'express';
-import WorkPackagesController from '../controllers/work-packages.controllers';
 import {
   descriptionBulletsValidators,
   intMinZero,
@@ -8,11 +7,12 @@ import {
   validateInputs
 } from '../utils/validation.utils';
 import { body } from 'express-validator';
+import WorkPackageTemplatesController from '../controllers/work-package-templates.controllers';
 
 const workPackageTemplatesRouter = express.Router();
 
-workPackageTemplatesRouter.get('/', WorkPackagesController.getAllWorkPackageTemplates);
-workPackageTemplatesRouter.get('/:workPackageTemplateId', WorkPackagesController.getSingleWorkPackageTemplate);
+workPackageTemplatesRouter.get('/', WorkPackageTemplatesController.getAllWorkPackageTemplates);
+workPackageTemplatesRouter.get('/:workPackageTemplateId', WorkPackageTemplatesController.getSingleWorkPackageTemplate);
 
 workPackageTemplatesRouter.post(
   '/:workpackageTemplateId/edit',
@@ -25,7 +25,7 @@ workPackageTemplatesRouter.post(
   nonEmptyString(body('workPackageName').optional()),
   ...descriptionBulletsValidators,
   validateInputs,
-  WorkPackagesController.editWorkPackageTemplate
+  WorkPackageTemplatesController.editWorkPackageTemplate
 );
 
 workPackageTemplatesRouter.post(
@@ -39,9 +39,12 @@ workPackageTemplatesRouter.post(
   nonEmptyString(body('blockedBy.*')),
   ...descriptionBulletsValidators,
   validateInputs,
-  WorkPackagesController.createWorkPackageTemplate
+  WorkPackageTemplatesController.createWorkPackageTemplate
 );
 
-workPackageTemplatesRouter.delete('/:workPackageTemplateId/delete', WorkPackagesController.deleteWorkPackageTemplate);
+workPackageTemplatesRouter.delete(
+  '/:workPackageTemplateId/delete',
+  WorkPackageTemplatesController.deleteWorkPackageTemplate
+);
 
 export default workPackageTemplatesRouter;
